refactor: 탭별로 컴포넌트 분리
