+++ src/policies.ts
@@ -37,7 +37,7 @@ const calculate_추천_월_납입_금액 = (targetSavingsAmount: number, savingP
   return Math.round(rawAmount / 1000) * 1000;
 };
 
-const filterRecommendedSavingsProduct = (savingsProductList: SavingsProduct[]) => {
+const getRecommendedSavingsProduct = (savingsProductList: SavingsProduct[]) => {
   return savingsProductList.sort((a, b) => b.annualRate - a.annualRate).slice(0, 2);
 };
 
@@ -47,5 +47,5 @@ export const savingsProductCalculatorPolicy = {
   calculate_목표_금액과의_차이,
   calculate_예상_수익_금액,
   calculate_추천_월_납입_금액,
-  filterRecommendedSavingsProduct,
+  getRecommendedSavingsProduct,
 };
